Precisely preserve the timestamps when stripping debugging symbols
The `--preserve-dates` flag for `strip` and `objcopy` only has whole
second resolution.  This meant that the modification time was being
truncated to an integer number of seconds.  Consequently, when
stripping a file, some of the dependencies of that file would appear
to be newer, and hence the file would be unnecessarily remade during a
second call to make.

Simple soultion: copy the timestamps to a separate file before we
strip and copy them back after.